refactor(telemetry): replace global telemetry client with context-based handling
